requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
creds==1.0.15
discord==2.3.2
discord.py==2.3.2
frozenlist==1.4.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
multidict==6.0.4
python-telegram-bot==20.6
regex==2023.10.3
requests==2.31.0
sniffio==1.3.0
urllib3==2.0.6
yarl==1.9.2
